requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
async-timeout==4.0.3
asyncio-dgram==2.1.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imgkit==1.2.3
libtmux==0.25.0
mcstatus==11.0.1
MinePI==0.5.0
mojang==1.1.0
multidict==6.0.4
Pillow==10.1.0
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.1.0
websocket-client==1.6.4
yarl==1.9.3
